+++ apps/extension/src/apps/content/App/Account/Filters/Filters.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import ContentTitleTop from 'libs/components/organisms/Notification/ContentTitleTop';
-import withTitle from 'src/app/hocs/withTitle';
+import withTitle from 'libs/hocs/withTitle';
 
 export const Filters = () => {
   const { t } = useTranslation();