+++ apps/extension/src/apps/content/App/Contribute/ContributeScreen/index.ts
@@ -1,4 +1,4 @@
-import withTitle from 'src/app/hocs/withTitle';
+import withTitle from 'libs/hocs/withTitle';
 import ContributeScreen from './ContributeScreen';
 
 export default withTitle('title.post_form')(ContributeScreen);