+++ apps/extension/src/apps/content/App/Question/PreviewScreen/withTitle.ts
@@ -1,4 +1,4 @@
-import withTitle from '../../../../../../../../src/app/hocs/withTitle';
+import withTitle from '../../../../../../../../libs/hocs/withTitle';
 import { PreviewScreenOwnProps } from './PreviewScreen';
 
 export default withTitle<PreviewScreenOwnProps>('title.preview');